Avoid crashing when QtQuick destroys our SG node and GL context
Keep the layers on the UI process but make sure that the backing
stores are destroyed both on the UI and Web processes.

Add a check for m_isPurging in CoordinatedGraphicsLayer::removeTile
to make sure that we don't tell the UI process to remove tiles that
it purged itself, in a similar way to what we do in
CompositingCoordinator::removeUpdateAtlas.

Also make sure that we wait for the next m_isWaitingForRenderer
after we've purged the backing stores in CoordinatedLayerTreeHost
to wait for a QtWebPageSGNode to be available like we do right after
a page's construction and avoid sending update that would be
ignored by an unactive CoordinatedGraphicsScene.

Task-number: QTBUG-40734
Change-Id: I059f1de2ce2a8d9d70ab2cd8c3252a469ebf036e
Reviewed-by: Michael Bruning <michael.bruning@digia.com>